Manage subscriptions (#1117)
* Copy `promiseSeq` function from script to server.lib.utils

* Add script for charging subscriptions

The script `charge_subscriptions.js` will replace the subscriptions
via stripe. It's meant to run at least once a day and process all the
active subscriptions due until the date it is run.

Handling of each order:

  Failure scenarios:

   * If processing an order fails, a retry counter is incremented and
     the `nextChargeDate` is bumped to run two days after the failure;
   * The user gets notified every time charging fails.
   * After `lib.subscriptions.MAX_RETRIES` attempts, the subscription
     is cancelled and the user notified;

  Success scenario:

   * The `nextChargeDate` & `currentPeriodStart` are bumped to the
     next month (after currentPeriodStart) and the user is notified;

Migration: Three new fields are created in the table Subscriptions:

 1. `nextChargeDate`
 2. `currentPeriodStart`
 3. `chargeRetryCount`

* Remove subscription handling from stripe

With the new charge_subscriptions.js script, the code that handles
subscriptions via stripe isn't necessary anymore.

This commit removes the creation of the stripe subscriptions and the
webhook that finished the subscription charge is now a no-op.

* Add script to move subscriptions out of stripe

* Add Histories table to Subscriptions

* Email report at the end of the charge_subscriptions script

* Fix tests requiring stripe base subscription data

* Remove more tests related to stripe based subscriptions

* Update test databases with Subscription migrations

* Remove stripe call from subscription cancellation

* Add text email template for failed payments (subscriptions)

* Only activate subscription after creating transaction instance

* Fix output of charge_subscriptions script when error contain quotes

* Enable charge_subscriptions script to find all order subscriptions

* Handle charge dates not migrated as well as unpaid & past_due

We want to continue to charge pre-migration folks the day they signed
up. Because of that, setting `nextChargeDate' and `nextPeriodStart' to
the first day of the next month should only happen when subscriptions
are being created.

The `charge_subscriptions' script will only increment the charge dates
by a month or year, it won't find the first day of that month anymore.

The unpaid & past_due subscriptions are going to be charged as soon as
the migration runs, however, their `nextPeriodStart' doesn't start
until the end of the current billing period from stripe.

* Add link to manage subscriptions in payment.failed email template

* Add field `interval' to CSV output of subscription migration script

* Error out when stripe sends a subscription event to our webhook

This will help detecting subscriptions that should have been migrated
by printing it out on to the error log.

* Move processOrderWithSubscription to lib/subscriptions & tests

* Temporarily disable sending emails if charges to a subscription fail