scripts/charge_subscriptions: Group processed orders by status (#1150)
Now instead of a single list of processed orders, the email will
contain different groups based on the status of the processing of each
order. The three possible groups are: 1. charged, 2. past_due,
3. canceled.

Each group now also shows a sum of the amount of their entries (not
accounting for different currencies)