Update charge subscriptions (#1163)
* Rename email template thankyou.sustain to thankyou.sustainoss

There's some logic in the email library to concatenate the name of the
Collective's slug to the template name if it starts with
`thankyou`. However, the email template was called `sustain` instead
of `sustainoss`, which caused the `charge_subscriptions` script to
blow up upon a successful charge.

* charge_subscriptions: Ensure subscription is always updated

There was an error sending a confirmation email after successfully
charging a subscription. That error caused the whole script to stop
and no subscriptions on that batch were updated. I manually checked
which subscriptions were charged on Stripe and updated the database
accordingly.

From now one this kind of error will be logged but it won't interrupt
the script anymore.

* charge_subscriptions: Replace `dryrun` with `notdryrun` flag

* charge_subscriptions: Move back to scripts folder