ledger_fixer: More accurate report of changes & fixes
Lots of operations were always considering both transactions of a
pair (CREDIT & DEBIT) but some errors only happen to one of them. The
output of the report was off because of this.

The script is also doing a better job at not really changing a value
unless it really needs to. Especially when the field already contain
the value that the script wants to write.

We're back to ignoring the transactions of different currencies for
now and the recalculation of amounts is back.