test(store): refactor creating collectives & add createExpense
For extending the functionality while keeping name consistency, the
function `collectiveWithHost` now just creates a collective and
associates it an already existing host. The new function
`hostAndCollective` is the utility for creating a collective with a
host at one shot (it also returns the admin user of the host).

The underlying GraphQL mutation `createExpense` was also added to the
store to avoid importing that huge path in every where we need to
create expenses.