chore(tests): Allow tests to create transactions in the past
The helper `stores.stripeOneTimeDonation` now accepts the optional
parameter `createdAt`. If it's present, it will freeze the date using
`sinon.useFakeTimers`, execute the order and then restore the time.

That's useful because a lot of tests are time dependent and with this
features, the tests can just create the transactions and query for
them with whatever date/time and that will just work.